Balance working

+++ shared/src/commonMain/kotlin/com/arnyminerz/filamagenta/cache/data/AccountTransactionUtils.kt
@@ -3,6 +3,7 @@ package com.arnyminerz.filamagenta.cache.data
 import com.arnyminerz.filamagenta.cache.AccountTransaction
 import com.arnyminerz.filamagenta.network.database.SqlTunnelEntry
 import com.arnyminerz.filamagenta.network.database.getDate
+import com.arnyminerz.filamagenta.network.database.getDouble
 import com.arnyminerz.filamagenta.network.database.getLong
 import com.arnyminerz.filamagenta.network.database.getString
 
@@ -12,7 +13,7 @@ fun List<SqlTunnelEntry>.toAccountTransaction(): AccountTransaction {
         date = getDate("Fecha")!!,
         description = getString("Concepto")!!,
         units = getLong("Unidades")!!,
-        cost = getLong("Precio")!!,
+        cost = getDouble("Precio")!!,
         income = getString("Tipo") == "I",
     )
 }
